requirements/dev.txt
-r base.txt
django-debug-toolbar
# Required by django-debug-toolbar
sqlparse

coverage
flake8
# Required by flake8
pyflakes
pep8
mccabe
